CBES-110 Need document routing to support join
Motivation
==========

Multiple users have expressed an interest in custom document routing,
even with the limitation that the connector is unable to delete
the documents that use custom routing.

Modifications
=============

Add a new `routing` field to the document type config. Value is a
JSON pointer to the field to use for routing (usually parent ID).

When indexing a document, look for the field and set the routing
accordingly. If the routing field is missing or invalid, log a warning
and fall back to default routing.

Enforce the constraint that `ignoreDeletes` must always be `true`
if `routing` is specified.

Event.getContent() now caches the byte array so it's inexpensive to
call multiple times.

Change-Id: I5dacd13b6a5c78840e2d49e826c74228eb3dfd53
Reviewed-on: http://review.couchbase.org/105624
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Nocentini <james.nocentini@gmail.com>