CBES-121 Need better metric for reporting replication progress
Motivation
----------
The 'cbes.backfill' and 'cbes.backfillRemaining' metrics are useless
after the initial backfill is processed.

Modifications
-------------
Add a 'cbes.backlog' gauge which dynamically calculates the
difference between the high sequence number in Couchbase and the
last sequence number processed by the connector. The reported value
is the sum of the differences for all partitions handled by the
connector instance.

Change-Id: I13c12a0dc3fcef5e8a8bc223e4b3bcbcde1b7b37
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/138798
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>