CBES-181 Add config option to disable TLS hostname verification
Motivation
----------
Hostname verification used to be disabled by default, but this changed
in connector version 4.2.3 as part of the upgrade to Couchbase SDK 3.

Some users are having trouble migrating to the latest connector because of
limitations in their certificate infrastructure/policies. Allowing users
to revert to the old bevahior (not verifiying hostnames) eases the migration.

Modifications
-------------
Add a `hostnameVerification` option to the `couchbase` config section.
Pass this value along to the Couchbase SDK.

Change-Id: Ic3559de0cd34c0aaea7a9ae535fadd692c9450f3
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/141397
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Smedley <richard.smedley@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>